Break out DataStoreEntry
